move pd scheduler to separated pkg (#532)
* move pd scheduler to separate struct

* add comments

* fix import

* fix test

* add NewPdController

* add missing comment

* move pd controller to separated pkg

* fix

* make PDController in Mgr anonymous
